finish merge and fix root tree malloc error

+++ src/apply_execution_function.c
@@ -27,7 +27,7 @@ struct Value apply_execution_function(struct Tree * ast, hash_table * defined, s
     }
     else if(ast->content.type == KEYWORD && string_matches_heap(&let_const, &ast->content.data.str)){
         if (ast->size < 2) {
-            ERROR("Not enough arguments for store_let_binding: %i < 2\n", ast->size);
+            ERROR("Not enough arguments for store_let_binding: %zu < 2\n", ast->size);
         }
         store_let_binding(ast->children[0],ast->children[1], defined, scopes, function_names, max_depth-1);
         return undefined;
@@ -48,7 +48,7 @@ struct Value apply_execution_function(struct Tree * ast, hash_table * defined, s
     }
     else if(ast->content.type == KEYWORD && string_matches_heap(&set_const, &ast->content.data.str)) {
         if (ast->size < 3) {
-            ERROR("Not enough arguments for set: %i < 3", ast->size);
+            ERROR("Not enough arguments for set: %zu < 3", ast->size);
         }
         struct Value index = execute(ast->children[0], defined, scopes, function_names, max_depth-1);
         struct Value item = execute(ast->children[1], defined, scopes, function_names, max_depth-1);
@@ -62,9 +62,9 @@ struct Value apply_execution_function(struct Tree * ast, hash_table * defined, s
     }
     else if(ast->content.type == KEYWORD && string_matches_heap(&do_const, &ast->content.data.str)){
         if (!ast->size) {
-            ERROR("Not enough arguments for do: %i < 1", ast->size);
+            ERROR("Not enough arguments for do: %zu < 1", ast->size);
         }
-        for(int i = 0; i < ast->size; i++){
+        for(size_t i = 0; i < ast->size; i++){
             if(i == ast->size-1){
                 return execute(ast->children[i], defined, scopes, function_names, max_depth-1);
             } else {
@@ -74,7 +74,7 @@ struct Value apply_execution_function(struct Tree * ast, hash_table * defined, s
     }
     else if(ast->content.type == KEYWORD && string_matches_heap(&read_const, &ast->content.data.str)){
         if (ast->size < 1) {
-            ERROR("Not enough arguments for read: %i < 1", ast->size);
+            ERROR("Not enough arguments for read: %zu < 1", ast->size);
         }
         char t = ast->children[0]->content.type;
         if (t != STRING && t != KEYWORD) {
@@ -84,12 +84,12 @@ struct Value apply_execution_function(struct Tree * ast, hash_table * defined, s
     }
     else if(ast->content.type == KEYWORD && string_matches_heap(&substring_const, &ast->content.data.str)){
         if (ast->size < 3) {
-            ERROR("Not enough arguments for substring: %i < 3", ast->size);
+            ERROR("Not enough arguments for substring: %zu < 3", ast->size);
         }
         struct Value string = execute(ast->children[2], defined, scopes, function_names, max_depth-1);
         struct Value start = execute(ast->children[0], defined, scopes, function_names, max_depth-1);
         struct Value end = execute(ast->children[1], defined, scopes, function_names, max_depth-1);
-        
+
         return substring(start, end, string);
     }
     else if(ast->content.type == KEYWORD && string_matches_heap(&switch_const, &ast->content.data.str)){
